"esqueleto" principal pronto, volume agora é float
gvt agora é vivo

+++ musicInterpreter/src/musicinterpreter/Nota.java
@@ -12,7 +12,7 @@ package musicinterpreter;
 public class Nota {
     private String nomeNota;
     private int oitava;
-    private int volume;
+    private float volume;
     
     public aumentaOitava(){
         if(oitava < 8)
